-- simple renaming of the test cases for consistency
