+++ tests/base/src/test/java/org/keycloak/tests/oid4vc/preauth/OID4VCSdJwtAuthorizationDetailsFlowPreAuthTest.java
@@ -31,7 +31,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  * SD-JWT-specific authorization_details tests for pre-authorized_code grant.
  */
 @KeycloakIntegrationTest(config = OID4VCIssuerTestBase.VCTestServerWithPreAuthCodeEnabled.class)
-public class OID4VCSdJwtAuthorizationDetailsFlowPreAuthTest extends OID4VCAuthorizationDetailsFlowTestBase {
+public class OID4VCSdJwtAuthorizationDetailsFlowPreAuthTest extends OID4VCAuthorizationDetailsFlowPreAuthTestBase {
 
     @Override
     protected String getCredentialFormat() {
